requirements.txt
rich~=14.3.2
pdfplumber~=0.11.9
langchain~=1.2.0
langchain-openai~=1.1.6
langchain-community~=0.4.1
langchain-core~=1.2.5
langchain-text-splitters~=1.1.0
openai~=2.14.0
faiss-cpu~=1.13.2
tiktoken~=0.12.0
python-dotenv~=1.2.1

